refactor: move provisioning package to pkg/internal
